Fix file sorting for more than 11 elements
V8's Array.prototype.sort uses combination of quick sort and insertion sort.

The compare function should return integer (-1, 0, 1), but we were returning a boolean, which is incorrect. Boolean true equals to 1, but boolean false equals to 0, not -1. The way the condition was written, it worked for small arrays (where only insertion sort, which is stable, was used), but not for larger arrays, when unstable quick sort screwed it up.