feat (reporter.stream): Add a reporter suitable for capture to a file
The project is moving to log4js for logging but while tests are running
as part of an ant build can can be useful to have a reporter that just
streams the output to the console.  Progress is okay at this, but also
spew navigation characters.  If you're looking the output from a text file,
it is very ugly.

The Stream reporter is a stripped down thin wrapper around Base. I'm not
sure it makes sense for it to have a color option, but I've included it
for now.