feat(config): `browsers` can specify args and cmd
Support optional `cmd` and `args` parameters in the `browsers`
configuration. The `args` are appended to the default command-line
options for the browser. If `overrideArgs` is true, then the `args`
replace the default command-line options for the browser. Occurrences
of `$TEMPDIR` or `$URL` within the `args` are replaced by the temp
directory and URL. Use case is to pass special options, e.g., to
load browser extensions.

Closes #278