fix(logger): configure the logger as soon as possible
This is a tricky problem - we need to show some logs even before parsing the configuration. At that point, we don't know what the logger configuration is.

This sets the reasonable defaults at the very beginning (rather than log4js verbose defaults) and then, sets the user specific settings as soon as possible.