feat(config): add usePolling config
This config option is forwarded to chokidar. When set to `true` (default value is `false`), chokidar uses `fs.watch` instead of stat polling.

This seems to be finally working well on linux, however it does not work on Mac.
See https://github.com/paulmillr/chokidar/issues/62

That's why I'm leaving it undocumented ;-)