fix(config): ignore empty string patterns
An empty string pattern is a mistake - there is no reason to do that. It can however happen (for instance when you generate the config file).

Before this fix, Karma would try to watch this empty string pattern, which would result in watching the entire `basePath` directory.

Now, the pattern will be ignored and a warning showed.