chore(deps): Remove peer dependencies.
Peer dependencies should not go from the parent to the children. The children
should peer depend on the parent project and not the other way around.

Closes #767

BREAKING CHANGE: Karma does not ship with any plugin. You need to explicitly install all the plugins you need. `karma init` can help with this.

Removed plugins that need to be installed explicitly are:

* karma-jasmine
* karma-requirejs
* karma-coffee-preprocessor
* karma-html2js-preprocessor
* karma-chrome-launcher
* karma-firefox-launcher
* karma-phantomjs-launcher
* karma-script-launcher