feat(web-server): disable gzip compression
With big files, the gzip compression is causing to serve only parts (chunks) of a file, as browser probably cancels the request at some timeout.

See https://github.com/karma-runner/karma/pull/778 for a discussion. I saw similar issues on AngularJS build couple of times too.

Doing gzip compression on the fly is not a good idea. The compress module is more likely to be removed from express/connect.

If we wanna do gzip compression, we should probably do it as a preprocessor, so that it happens only once per file.