fix(web-server): detach listeners after running
Remove listeners from the global process and the webserver when we're
done. The EmitterWrapper can be used to wrap any EventEmitter and
remembers the listeners that were via the wrappers addListener/on
function. Then the wrapper's removeAllListeners function can be used
to remove the wrapper's listeners, leaving the listeners that were
added directly to the unwrapped emitter attached.