feat(server): wait for frameworks to load via q.all() before starting server
The API is to simply return a promise from a factory function, or use a promise as a value. The server
will wait for the collection of frameworks to resolve before carrying on starting the server.

This is useful for when setup requires some asynchronous startup.

Currently, any Promise api which is interoperable with q may be used.

In addition to this feature, this CL also adds a set of callbacks to the server, via an object for the
3rd parameter. This is useful for test purposes, and enables the test case to run faster, by not waiting
for the server to shut down via the second parameter.

Closes #851