chore(presubmit): make master stable, use canary branch
So far it looks like most of the PRs are fixes should be based on “stable” branch. That is however
not intuitive as by default, a branch is started from “master”. I believe using “master” as the
stable branch and “canary” branch as the one that gets pushed into “canary” NPM channel (and
contains upcoming feature release) is better.