fix(server): launch browsers when file_list is ready
If we have a lot of files and a fast starting browser (PhantomJS),
the browser might try to connect to the web server before the
server is ready to reply, causing the connect mechanism to time out.
Make sure to wait until fileList.refresh() is complete before starting
the browsers.