feat(runner): Use favicon in static runner pages
I noticed a "TOOD" [sic] in the runner page source that was added by
e9fa51ac58daaf8. I figured I'd take a swing at doing it. This commit
brings down the favicon from

  http://karma-runner.github.io/assets/img/favicon/favicon.ico

and wires up the static pages to use it. I considered inlining it as a
data-uri, but the image was big enough to make the source pretty gross
so I settled on using an external file.

Along the way, I found and fixed a couple of minor spelling and
whitespace errors.