feat(config): add nocache option for file patterns
Add nocache property for file patterns in the config file.
This allows files that should be served by the webserver but that
don't need to be preprocessed but that should to be served fresh
(e.g. map files, pre-transpiled source built by other build process)
without a large amount of watches.