fix(launcher): Continue with exit when SIGKILL fails
We have seen very occasional hangs where the browser fails to capture and then
also fails to exit on SIGKILL.  This results in karma hanging waiting on the
browser to restart.  It is not great to leave zombie processes around, but it is
worse to hang.

This change makes karma continue if SIGKILL doesn't make a spawned process
exit in reasonable time, same as the regular killTimeout.