fix(server): Start webserver and browsers after preprocessing completed
When loading many files, it is possible that the webserver and browsers
will start before file preprocessing has completed. This causes the
browser to attempt to connect and timeout and need to be relaunched.

This change will cause the server to wait to start the webserver and
browsers until after all files have been preprocessed. When a browser
connects to the webserver, the webserver will be able to start serving content
immediately and the browser will not experience any connection errors or
timeouts.