fix (reporter): better handling of stack trace source map rewrite
In some cases, reported stack traces do not contain the url scheme and
domain. This is the case when using the webpack preprocessor where stack
entries look like: `at base/main.js?<sha1>:44977`. The previous url regexp
was always assuming an url scheme and domain.

This commit make the url scheme and domain optional so more stack traces
can be handled when trying to link from a transformed source code to the
original location provided by the source map associated with the served
file.