fix(web-server): Correctly update filesPromise on files updated
When the `file_list_modified` event was being emitted, the filesDeferred
was simply being resolved again; this has no effect and subsequent
calls to `filesPromise.then` continued to be resolved with the old
files.

This had the effect that issuing `karma run` did not result in new test
files being served.

This has been fixed by reverting the changes to `web-server.js` made in
fabdbae.