chore: Tweak supported Node.js versions, drop io.js
Current setup means karma won't be tested on future updates to the Node v4
line. Those updates are not allowed to introduce breaking changes and only
the newest one is officially supported so that's the one that should be tested.

In the same way, ythe config doesn't contain "0.10.0" or "0.12.0" but
"0.10" & "0.12" so it catches latest updates.

io.js testing has been dropped - io.js has no official support policy as
opposed to Node.js; also, it won't see any major releases.

Refs 6b8d30f234aa77c87ac8cd720114c90a66e7e88e