chore(travis): Test on Node 5, use default npm; uprade g++
Node 5 includes npm 3 by default so we don't have to update npm to make sure
it works on that version. On the other hand, most people don't upgrade npm
and just use the default included version, especially that it's very hard
to update npm on Windows. It's good to make sure Karma still works with npm 2,
especially that Node 4, current LTS will never upgrade to npm 3.

g++ got upgraded so that native module compilation works on Node >= 4;
otherwise lots of optional packages fail, making the log too large for Travis
which in turn kills the job.