chore(package): add es3 babel preset
Adding the es3 babel preset ensures the highest compatibility with
browsers like IE8.
In the babelrc file, the preset comes before the es2015 preset because
the presets are executed in reverse order. So, the es2015 preset will
run before the es3 preset, which is what we want to have happen.