fix(file-list): always use file from first matcher
If a file is matched by multiple patterns, only the file from
the first pattern should be used.

This fixes issues that occurr when the same file appears in
multiple buckets. One such issue is that the watcher reruns
the preprocessors on the first occurrence of the file but
the second version of the file is the one being included in
the context. This results in the browser not requesting the
updated version of the file after the watch fires.