Add a little paragraph about scopes
This is a proposal to explain a little bit more how to use scopes.

This is how I understood them but I would love to have your opinions on the different ways to use scope.

Should it be a real list across every projects? Project-specific each time? Should it talk about user-centric scopes or technical ones?