docs(config): add reminder about relation between `runInParent` and watch mode
Learned the hard way that running the tests in the same window as the client prevents the server from persisting and re-running on file change. Makes sense on a top-level because it requires a separate window or iframe to do the watch, but figure it be a good place to be redundant about this tip.