refactor: subscribe to socket events explicitly
This makes code easier to understand, because all listeners are added
explicitly. Now it is clear what methods are called in response to which
events and what parameters are passed to them.

Also it is a preparation for migration to ES2015 classes, where original
approach won't work, because methods in ES2015 classes are added as not
enumerable properties, so it is harder to loop through them.