refactor(server): Provide file-service handlers in the root injector. (#3042)
By removing the creation of file-service handlers from the webserver, we create functions with better focus
and we allow the file-service handlers to be used in beforeMiddlewaare. The only cost is that these objects are
now visible to all modules rather than being private to webserver. The potential for collision seems small.