fix: remove circular reference in Browser
After migrating the the `browser.js` file to ES2015, the `Browser` class has now fields with circular references (like `emitter`),
which breaks the Karma JSON reporters available starting with Karma 2.0.1.
This commit removes the circular dependencies by adding a `toJSON()` method to the `Browser` class,
which produces a similar result to the browser object before the migration.

Fixes #3075