fix(browser): ensure browser state is EXECUTING when tests start (#3074)
* fix(browser): ensure browser state is EXECUTING when tests start

Browser state is EXECUTING when it actually started to execute tests. This state change is triggered by client on actual tests execution start.

Introduced an additional browser state CONFIGURING

The CONFIGURING state means that the browser is not just CONNECTED for tests, but someone has requested tests execution (and provided a config file). But the provided config file is not yet processed, configuration is not applied or the tests execution is not yet started and we have not received the first event from the remote browser, so the browser object is not yet at EXECUTING state.

Refactored browser state names: renamed READY -> CONNECTED

Fixes #1640