chore(test): fix flaky test cases (#3314)
* fix(test): fixed flaky reconnecting test

https://github.com/karma-runner/karma/commit/cc2eff27deb680f789afb34577fd337d2ad5dcac introduced different behavior depending on whether socket reconnected or browser reconnected after reload. This lead to the reconnecting test being flaky as, depending on the timing, it will trigger second test run and result in unexpected output.

* fix(test): remove broken monitor action for Cucumber

The issue is that it resulted in the following sequence of actions:
- run karma start
- run karma run
- and then instantly kill karma start command

This resulted in flaky tests because the kill could race, resulting in
incomplete output. In fact test was unnecessary complicated as it was
enough to use runOut, which asserts output of the karma run command.

* fix(test): guard from repeated executions of the karma run

'data' event handler may be called multiple times, which will result in
multiple calls to karma run. To take it even further, its execution is
delayed by setTimeout, which means that subsequent executions will run
while next test scenario is in progress and may mess it up. To prevent
this make sure that karma run is executed only once independently of how
many time 'data' event is fired.