chore(build): run cucumber tests using NPM scripts instead of Grunt
The main reason for this change is to be able to update Cucumber library as `grunt-cucumberjs` (https://github.com/mavdi/grunt-cucumberjs/issues/110) does not support cucumber@4 and looks pretty abandoned.

Grunt ecosystem is not very active for a while, so we can expect more issues of this sort in the future. Because of this and to keep docs consistent I've also added NPM scripts to run other test types (for now they just proxy to Grunt) and removed Grunt examples from the documentation.

In the next PR I will update `cucumber` and perform some cleanup of the step definitions.