refactor(test): add common method to start server in background
The existing code had pretty confusing logic (old lines 33-36)`: when background server start fails it is actually considered a success, but child process is not started. Because of this `lastRun` contains result output of the `karma start`. This is no longer the case, hence two test cases had to be updated as they never executed `karma run` despite the statement and were asserting against `karma start` output. This should be more clear now.

As background server process now has its own variable to store output, there is no need for a dedicated runOut command and it can be removed.