chore(test): remove @not-jenkins tag for Cucumber tests (#3479)
This tag was useful long time ago, when most of the tests were run using PhantomJS. For a long time E2E tests use ChromeHeadless as a the main browser and it is not needed to mark tests using Google Chrome to only run locally. Two removed test cases have identical cases above them, which are run in the ChromeHeadless and other two were updated to run in ChromeHeadless instead of Chrome and will be run both in CI as well as locally.