chore(test): fix client tests in older IE versions
By locking sinon to the older version. The problem was that new version of sinon brought newer versions of several dependencies, which use ES2016 Map, which is not available in the older IE versions. The solution is fragile as we can't reliably lock transitive dependencies, but should do for now to make master green.

As a proper solution we should consider cutting a new major release, where we would drop support for ancient browsers and more clearly document which browsers karma aims to support.