feat(bundler): add a bundler layer
The input config supplies 'patterns' which FileList expands into Files.
These files are preprocessed and then 'file-list-modified' is fired.
The event offers a pair of file lists, {included, served}.
The bundle layer maps this pair to a new pair where fewer files are included.
By default the bundler copies the inputs to outputs.

The bundled files are stored on currentWebFiles which replaces
the FileList.files in the middleware/karma.