docs: fix weird syntax to reduce diff
Old parser considered this the same list, but new parser assumes that these are two separate lists. Technically it should be a nested list, but it is old and shouldn't really matter. I'm doing this change to reduce the diff I need to review after switching to the new markdown parser.