docs: updates to the documentation to support new markdown renderer (#3672)
Companion PR for https://github.com/karma-runner/karma-runner.github.com/pull/62.

* docs: fix invalid heading syntax

It is not recognized as a heading by the markdown parser without a whitespace.

* docs: fix weird syntax to reduce diff

Old parser considered this the same list, but new parser assumes that these are two separate lists. Technically it should be a nested list, but it is old and shouldn't really matter. I'm doing this change to reduce the diff I need to review after switching to the new markdown parser.

* docs: fix invalid markdown

Items nested in the list item, must be indented 3 spaces, not 2 spaces. New parser produced a pretty messed up markup from the invalid markdown.

* docs: remove redundant whitespaces in the list items