chore(ci/client/browserstack): modify concurrency and timeout settings
In our client tests, for BrowserStack, reduce the `browserDisconnectTimeout`
and the `browserNoActivityTimeout` to a minute, instead of 5. Also, reduce the
concurrency of browsers to 2 instead of 3. This is because BS is being flaky
when handling many browsers concurrently and, when a browser has been
disconnected it will generally not come back, so we do not need to wait a whole
5 minutes; it is better to just use another try from the
`browserDisconnectTolerance` configuration pool. The overall result is a more
efficient CI run.