fix(middleware): replace %X_UA_COMPATIBLE% marker anywhere in the file
Previously %X_UA_COMPATIBLE% marker was only replaced if it was located at the start of the line.
The limitation looks pretty arbitrary and caused the marker not to be replaced in the custom
debug.html file used by Angular CLI as the marker was not located at the start of the line
(probably because the file was re-formatted). This commit changes the behavior to replace the
marker anywhere within the file, not just at the start of the line and thus fixes the problem
for Angular CLI and potentially other people using custom files.

Fixes #3711