fix(helper): make mkdirIfNotExists helper resilient to concurrent calls
The main motivation for this change is https://github.com/karma-runner/karma-coverage/issues/434#issuecomment-1017939333 where concurrent calls to the helper fail because of the race between the check and directory creation. This is a temporary solution and should be replaced with the native [`mkdir`](https://nodejs.org/api/fs.html#fsmkdirpath-options-callback) with the `recursive` option once minimum supported Node is bumped to 12.