ci: prevent duplicate CI tasks on creating a PR
BrowserStack can be flaky when there are multiple concurrent jobs running on
it. This commit makes sure that only 1 browser can run concurrently during the
`npm run test:client` jobs and that the "Test" GitHub Action workflow is only
trigerred once when creating a PR; having the `on: push` configuration made it
so that GH triggers two duplicate jobs when a PR is opened.